requirements_fastvlm_0.1.0.txt
accelerate==1.6.0
aiofiles==23.2.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bitsandbytes==0.48.1
blinker==1.9.0
cattrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
coremltools==8.2
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
fastapi==0.120.0
ffmpy==0.6.4
filelock==3.20.0
Flask==3.1.2
fsspec==2025.9.0
gradio==5.11.0
gradio_client==1.5.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
latex2mathml==3.78.1
lazy_loader==0.4
-e git+https://github.com/apple/ml-fastvlm.git@592b4add3c1c8a518e77d95dc6248e76c1dd591f#egg=llava
loguru==0.7.3
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.8.1.78
orjson==3.11.4
packaging==25.0
pandas==2.3.3
peft==0.13.2
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.0
psutil==7.1.1
pyaml==25.7.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pynvml==13.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rich==14.2.0
ruff==0.14.2
safehttpx==0.1.7
safetensors==0.6.2
scenedetect==0.6.7.1
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.0
tomli==2.3.0
tomlkit==0.13.3
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wavedrom==2.0.3.post3
websockets==14.2
Werkzeug==3.1.3
google-cloud-storage==2.12.0